+++ sentry_link/lib/sentry_link.dart
@@ -5,3 +5,4 @@ export 'src/sentry_tracing_link.dart';
 export 'src/gql_event_processor.dart';
 export 'src/sentry_request_serializer.dart';
 export 'src/sentry_response_parser.dart';
+export 'src/sentry_breadcrumb_link.dart';